Move RunInstallers to InitializableEndpoint.Initialize()
